requirements.txt
bs4
requests
selenium
matplotlib
seaborn
plotly
pyyaml